Relocate fangraphs.leaders.ScrapingUtilities to fangraphs.ScrapingUtilities
